requirements.txt
pyyaml
aiohttp
cryptography
packaging